Convert Fixed & Modal components to tsx
See: #44
Change-Type: patch
